frontend/scripts/logo-downloader/requirements.txt
# Dépendances pour download_logos.py
#
# Installation :
# pip install -r requirements.txt

requests>=2.31.0
Pillow>=10.0.0
